Ensure invoices show most recently due first, not oldest
When a client has dozens of invoices, it doesn't make any sense to show the oldest ones first (after sorting by due date that is). This means that if all invoices are paid, it might show invoices from years ago rather than the most recent ones.

This change ensures that unpaid invoices are shown first, then the most recent invoices.